Created model for activities with proper types.
Implemented listing API for activities with funtionality to read all of them when the API is called.

Generated activity on expense create for included members only.

Generated activity on expense update when either expense amount is changed, splits are changed, a member is added in split, a member is removed from split, item's amount is changed, any item is deleted.

Generated activity on add group member.

Added has_unread_activities boolean attribute in login and me endpoints to indicate if the user has unread activities.